requirements.txt
Django>=4.2.7
Pillow>=10.1.0
django-modeltranslation>=0.18.11
django-ckeditor>=6.7.0
gunicorn>=21.2.0
